+++ config/initializers/content_security_policy.rb
@@ -11,14 +11,11 @@ Rails.application.configure do
     policy.img_src     :self, :https, :data
     policy.object_src  :none
     # policy.script_src  :self, :https
-    policy.script_src  :self, :https, :unsafe_inline
+    policy.script_src  :self, :https, :unsafe_inline, :blob
     # policy.style_src   :self, :https
     policy.style_src   :self, :https, :unsafe_inline
     # Specify URI for violation reports
     # policy.report_uri "/csp-violation-report-endpoint"
-    # :nocov:
-    policy.connect_src :self, :https, 'http://localhost:3035', 'ws://localhost:3035' if Rails.env.development?
-    # :nocov:
   end
 
   # Generate session nonces for permitted importmap and inline scripts