+++ app/models/application_record.rb
@@ -1,5 +1,5 @@
 class ApplicationRecord < ActiveRecord::Base
-  self.abstract_class = true
+  primary_abstract_class
   has_paper_trail
 
   def self.search_like