+++ config/initializers/rails_admin.rb
@@ -3,7 +3,7 @@ require 'nested_form/engine'
 require 'nested_form/builder_mixin'
 
 RailsAdmin.config do |config|
-  config.asset_source = :webpacker
+  config.asset_source = :sprockets
   ### Popular gems integration
 
   ## == Devise ==