+++ config/routes.rb
@@ -1,4 +1,9 @@
 Rails.application.routes.draw do
+  # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
+  #
+  # Defines the root path route ("/")
+  # root "articles#index"
+
   draw :downloads
   draw :members
   draw :invitations
@@ -8,6 +13,7 @@ Rails.application.routes.draw do
   draw :admin
   draw :users
   root 'top#index'
+
   # :nocov:
   mount LetterOpenerWeb::Engine => '/letter_opener' if Rails.env.development?
   # :nocov: