+++ spec/support/application_contexts.rb
@@ -63,10 +63,12 @@ def expect_image_json(response_json_model, model)
   expect(data.count).to eq(5)
 end
 
-def get_locale(key, **replace)
-  result = I18n.t(key, **replace)
+def get_locale(key, **)
+  result = I18n.t(key, **)
+  # :nocov:
   raise if /translation missing:/.match(result)
 
+  # :nocov:
   result
 end
 
@@ -209,8 +211,10 @@ shared_examples_for 'ToNG(json/json)' do |code, errors, alert = nil, notice = ni
   end
 end
 shared_examples_for 'ToNG(html)' do |code, errors = nil|
+  # :nocov:
   raise 'errors blank.' if code == 422 && errors.blank?
 
+  # :nocov:
   let(:subject_page) { 1 }
   it_behaves_like 'ToNG(html/html)', code, errors
   it_behaves_like 'ToNG(html/json)', code