+++ RELEASES.md
@@ -2,7 +2,7 @@
 
 ## 1.8.0 (Sprint 13)
 
-- Ruby3.2.2/Rails7.0とGem/Packageをバージョンアップしたい
+- Ruby3.2.2/Rails7.0（WebpackerからSprocketsに移行）とGemをバージョンアップしたい
   - https://dev.azure.com/nightonly/rails-app-origin/_workitems/edit/557/
 
 ## 1.6.0 (Sprint 11)